+++ openbas-model/src/main/java/io/openbas/database/raw/impl/SimpleRawInjectExpectation.java
@@ -10,8 +10,8 @@ public class SimpleRawInjectExpectation implements RawInjectExpectation {
 
     private String inject_expectation_id;
     private String inject_expectation_type;
-    private Integer inject_expectation_score;
-    private Integer inject_expectation_expected_score;
+    private Double inject_expectation_score;
+    private Double inject_expectation_expected_score;
     private String team_id;
     private String asset_id;
     private String asset_group_id;
